[Update] DroneControlLibrary, drone_map [Update] Normalize latitude and longitude. (long -> float) [Add] GPSLock

+++ app/src/main/java/com/coretronic/drone/missionplan/fragments/HistoryFragment.java
@@ -69,7 +69,7 @@ public class HistoryFragment extends MapChildFragment {
 
             @Override
             public void onFlightHistorySelected(FlightHistory flightHistory) {
-                List<Long> flightRecords = new ArrayList<>();
+                List<Float> flightRecords = new ArrayList<>();
 
                 for (RecordItem recordItem : flightHistory.getRecordItems()) {
                     flightRecords.add(recordItem.getLatitude());
